refactor(cached_object_store): Move FsCacheStorage into a new module.
This refactor moves the code related to FsCacheStorage into a new module to make the code easier to extend with other backends, like RocksDB in the future. The change should be backwards compatible since it doesn't really change any functionality.
